requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
Django==6.0.3
django-allauth==65.15.1
dotenv==0.9.9
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
Jinja2==3.1.6
Levenshtein==0.27.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pillow==12.1.1
pycparser==3.0
Pygments==2.19.2
PyJWT==2.12.1
pyTelegramBotAPI==4.33.0
python-barcode==0.16.1
python-dotenv==1.2.2
python-Levenshtein==0.27.3
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
setuptools==68.2.2
shellingham==1.5.4
sqlparse==0.5.5
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wheel==0.46.3
